requirements.txt
fastapi
uvicorn
python-multipart
requests
openbharatocr
pydantic
tensorflow-cpu
numpy
opencv-python-headless
pillow
python-dateutil 
transformers>=4.23.0
torch
sentencepiece
